+++ gax-httpjson/src/main/java/com/google/api/gax/httpjson/longrunning/stub/OperationsStub.java
@@ -32,6 +32,7 @@ package com.google.api.gax.httpjson.longrunning.stub;
 import com.google.api.core.BetaApi;
 import com.google.api.gax.core.BackgroundResource;
 import com.google.api.gax.httpjson.longrunning.OperationsClient.ListOperationsPagedResponse;
+import com.google.api.gax.rpc.LongRunningClient;
 import com.google.api.gax.rpc.UnaryCallable;
 import com.google.longrunning.CancelOperationRequest;
 import com.google.longrunning.DeleteOperationRequest;
@@ -40,7 +41,6 @@ import com.google.longrunning.ListOperationsRequest;
 import com.google.longrunning.ListOperationsResponse;
 import com.google.longrunning.Operation;
 import com.google.protobuf.Empty;
-import javax.annotation.Generated;
 
 // AUTO-GENERATED DOCUMENTATION AND CLASS.
 /**
@@ -48,7 +48,6 @@ import javax.annotation.Generated;
  *
  * <p>This class is for advanced usage and reflects the underlying API directly.
  */
-@Generated("by gapic-generator-java")
 @BetaApi
 public abstract class OperationsStub implements BackgroundResource {
 
@@ -71,6 +70,10 @@ public abstract class OperationsStub implements BackgroundResource {
 
   public UnaryCallable<CancelOperationRequest, Empty> cancelOperationCallable() {
     throw new UnsupportedOperationException("Not implemented: cancelOperationCallable()");
+  }
+
+  public LongRunningClient longRunningClient() {
+    throw new UnsupportedOperationException("Not implemented: longRunningClient()");
   }
 
   @Override